split register actions into different files
